Update applicant data

+++ modules/backend-usermgt/backend-usermgt-service/src/main/java/org/opencps/usermgt/service/impl/EmployeeLocalServiceImpl.java
@@ -1196,11 +1196,18 @@ public class EmployeeLocalServiceImpl extends EmployeeLocalServiceBaseImpl {
 		// }
 		// }
 		// _log.info("user: "+JSONFactoryUtil.looseSerialize(user));
-		if (object.getWorkingStatus() == EmployeeTerm.WORKING_STATUS_RETIRED) {
-			userLocalService.deleteGroupUser(object.getGroupId(), object.getMappingUserId());
-		}
-		else if (object.getWorkingStatus() == EmployeeTerm.WORKING_STATUS_WORKED) {
-			userLocalService.addGroupUser(object.getGroupId(), object.getMappingUserId());
+		User u = userLocalService.fetchUser(object.getMappingUserId());
+		if (u != null) {
+			if (object.getWorkingStatus() == EmployeeTerm.WORKING_STATUS_RETIRED) {
+				if (object.getMappingUserId() > 0) {
+					userLocalService.deleteGroupUser(object.getGroupId(), object.getMappingUserId());
+				}
+			}
+			else if (object.getWorkingStatus() == EmployeeTerm.WORKING_STATUS_WORKED) {
+				if (object.getMappingUserId() > 0) {
+					userLocalService.addGroupUser(object.getGroupId(), object.getMappingUserId());
+				}
+			}
 		}
 		return employeePersistence.update(object);
